+++ admin/AppModel.ts
@@ -12,7 +12,7 @@ import {Route} from 'router5';
 import {activityTab} from './tabs/activity/ActivityTab';
 import {generalTab} from './tabs/general/GeneralTab';
 import {monitorTab} from './tabs/monitor/MonitorTab';
-import {serverTab} from './tabs/server/ServerTab';
+import {clusterTab} from '@xh/hoist/admin/tabs/cluster/ClusterTab';
 import {userDataTab} from './tabs/userData/UserDataTab';
 
 export class AppModel extends HoistAppModel {
@@ -63,9 +63,24 @@ export class AppModel extends HoistAppModel {
                 children: [
                     {name: 'about', path: '/about'},
                     {name: 'config', path: '/config'},
+                    {name: 'logLevels', path: '/logLevels'},
                     {name: 'users', path: '/users'},
                     {name: 'roles', path: '/roles'},
                     {name: 'alertBanner', path: '/alertBanner'}
+                ]
+            },
+            {
+                name: 'cluster',
+                path: '/cluster',
+                children: [
+                    {name: 'logs', path: '/logs'},
+                    {name: 'memory', path: '/memory'},
+                    {name: 'jdbcPool', path: '/jdbcPool'},
+                    {name: 'environment', path: '/environment'},
+                    {name: 'services', path: '/services'},
+                    {name: 'objects', path: '/objects'},
+                    {name: 'hibernate', path: '/hibernate'},
+                    {name: 'webSockets', path: '/webSockets'}
                 ]
             },
             {
@@ -75,20 +90,6 @@ export class AppModel extends HoistAppModel {
                     {name: 'tracking', path: '/tracking'},
                     {name: 'clientErrors', path: '/clientErrors'},
                     {name: 'feedback', path: '/feedback'}
-                ]
-            },
-            {
-                name: 'server',
-                path: '/server',
-                children: [
-                    {name: 'logViewer', path: '/logViewer'},
-                    {name: 'logLevels', path: '/logLevels'},
-                    {name: 'memory', path: '/memory'},
-                    {name: 'connectionPool', path: '/connectionPool'},
-                    {name: 'environment', path: '/environment'},
-                    {name: 'services', path: '/services'},
-                    {name: 'ehCache', path: '/ehCache'},
-                    {name: 'webSockets', path: '/webSockets'}
                 ]
             },
             {
@@ -119,14 +120,14 @@ export class AppModel extends HoistAppModel {
                 content: generalTab
             },
             {
+                id: 'cluster',
+                icon: Icon.server(),
+                content: clusterTab
+            },
+            {
                 id: 'activity',
                 icon: Icon.analytics(),
                 content: activityTab
-            },
-            {
-                id: 'server',
-                icon: Icon.server(),
-                content: serverTab
             },
             {
                 id: 'monitor',