+++ admin/tabs/cluster/websocket/WebSocketPanel.ts
@@ -4,18 +4,16 @@
  *
  * Copyright © 2024 Extremely Heavy Industries Inc.
  */
-import {WebSocketModel} from '@xh/hoist/admin/tabs/server/websocket/WebSocketModel';
+import {WebSocketModel} from '@xh/hoist/admin/tabs/cluster/websocket/WebSocketModel';
 import {grid, gridCountLabel} from '@xh/hoist/cmp/grid';
-import {filler, box, fragment, p} from '@xh/hoist/cmp/layout';
+import {filler, box, fragment, p, span} from '@xh/hoist/cmp/layout';
 import {relativeTimestamp} from '@xh/hoist/cmp/relativetimestamp';
 import {storeFilterField} from '@xh/hoist/cmp/store';
 import {XH, creates, hoistCmp} from '@xh/hoist/core';
-import {button, exportButton} from '@xh/hoist/desktop/cmp/button';
+import {exportButton} from '@xh/hoist/desktop/cmp/button';
 import {panel} from '@xh/hoist/desktop/cmp/panel';
 import {toolbarSep} from '@xh/hoist/desktop/cmp/toolbar';
-import {Icon} from '@xh/hoist/icon';
 import {errorMessage} from '@xh/hoist/desktop/cmp/error';
-import {AppModel} from '@xh/hoist/admin/AppModel';
 
 export const webSocketPanel = hoistCmp.factory({
     model: creates(WebSocketModel),
@@ -25,16 +23,12 @@ export const webSocketPanel = hoistCmp.factory({
 
         return panel({
             tbar: [
-                button({
-                    text: 'Force suspend',
-                    icon: Icon.stopCircle(),
-                    intent: 'danger',
-                    disabled: !model.gridModel.hasSelection,
-                    omit: AppModel.readonly,
-                    onClick: () => model.forceSuspendOnSelectedAsync()
+                span({
+                    item: 'WebSocket Connections',
+                    className: 'xh-bold'
                 }),
                 filler(),
-                relativeTimestamp({bind: 'lastRefresh'}),
+                relativeTimestamp({bind: 'lastRefresh', options: {prefix: 'Refreshed'}}),
                 toolbarSep(),
                 gridCountLabel({unit: 'client'}),
                 toolbarSep(),