+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## 64.0.0-SNAPSHOT - unreleased
 
+### 🎁 New Features
+* Provides admin support for Cluster-aware version of Hoist.
+
+### 💥 Breaking Changes
+* Requires update to `hoist-core >= 20.0.0`.
+
 ## 63.0.1 - 2024-04-05
 
 ### 🐞 Bug Fixes
@@ -9,6 +15,7 @@
 * New filterable fields exposed in the Admin Console for `ActivityTracking` and `ClientErrors` modules.
     * `url`, `appEnvironment`, `appVersion` in `ActivityTracking`
     * `impersonating` in `ClientErrors`
+    * 
 
 ## 63.0.0 - 2024-04-04